+++ markdown/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "module": "esnext",
     "target": "es2019",
     "lib": ["esnext"],
     "skipLibCheck": true,
